// frontend/src/client/types.gen.ts
export type ItemPublic = {
    title: string;
    description?: (string | null);
    filename: string;
    file_path: string;
    file_size: number;
    mime_type: string;
    width?: (number | null);
    height?: (number | null);
    alt_text?: (string | null);
    id: number;
    owner_id: number;
    collection_id: number;
    upload_date: string;
};